requirements.txt
requests
hana-ml
hdbcli
generative-ai-hub-sdk[all]
cryptography
pypdf
langchain-hana
ipykernel 
streamlit
langgraph
langchain
langsmith
python-dotenv
langchain_tavily



python-dotenv
sap-ai-sdk-gen[all] 
tqdm 
wikipedia